chore: remove unnecessary class wrapper over client
